refactor(internal): restructure JNI bindings; unified file and folder structure to align android and ios
