+++ ios/model_detect/sherpa-onnx-model-detect-helper.h
